training/distributed_training/pytorch/model_parallel/vit/requirements.txt
transformers==v4.21.0
evaluate
datasets
numpy
sagemaker
sagemaker-experiments
scipy
torchnet
smdebug
humanize